Create AbstractFilter for contains general actions.
