+++ wicket-orientdb/src/main/java/ru/ydn/wicket/wicketorientdb/utils/query/filter/number/EqualsNumberFilter.java
@@ -1,34 +1,27 @@
 package ru.ydn.wicket.wicketorientdb.utils.query.filter.number;
 
-import ru.ydn.wicket.wicketorientdb.utils.query.filter.IFilter;
+import ru.ydn.wicket.wicketorientdb.utils.query.filter.AbstractFilter;
 
 /**
  * value - value for equals
  * SELECT FROM class WHERE num = value
  */
-public class EqualsNumberFilter implements IFilter {
+public class EqualsNumberFilter extends AbstractFilter {
 
-    private final String field;
     private final Integer value;
-    private boolean join;
 
     public EqualsNumberFilter(String field, Integer value, boolean join) {
-        this.field = field;
+        super(field, join);
         this.value = value;
-        this.join = join;
+
     }
 
     @Override
-    public String apply() {
+    public String apply(String field) {
         StringBuilder sb = new StringBuilder();
         sb.append(field);
-        if (join) {
-            sb.append(" = ");
-            sb.append(value.toString());
-        } else {
-            sb.append(" <> ");
-            sb.append(value.toString());
-        }
+        sb.append(" = ");
+        sb.append(value.toString());
         return sb.toString();
     }
